matrix creation

+++ types.h
@@ -5,7 +5,7 @@
 
 typedef struct
 {
-    uint32_t *values;  // element values
+    double *values;  // element values
     uint32_t *row_idx; // cumulative elements of each row -> last elements equals NNZ
     uint32_t *col_idx; // column indices of the elements 
     uint32_t size;      // Matrix size (assuming square matrices only)
